requirements-dev.txt
ansible==4.3.0
docker==6.1.3
docker-compose==1.29.2
pytest==7.2.0
pytest-django==4.5.2
wheel==0.38.1
